pyproject.toml
[tool.poetry]
name = "tedis"
version = "0.1.0"
description = ""
authors = ["Lokrip <165576212+Lokrip@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
django-countries = "^7.6.1"
django-split-settings = "^1.3.2"
pyyaml = "^6.0.2"
pillow = "^11.1.0"
django-cors-headers = "^4.6.0"
django-mptt = "^0.16.0"
djangorestframework-simplejwt = "^5.4.0"


[tool.poetry.group.dev.dependencies]
colorlog = "^6.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
